Update: Score Input Screen Pending thorough testing.

+++ lib/app/modules/SetupScreen/competition_request_model.dart
@@ -1,7 +1,7 @@
 class CompetitionRequest {
   String compName;
   int gametypeId;
-  int compFee;
+  double compFee;
   String compDate;
   String compTime;
   String gameHoles;
